Update ESP32 to build with IDF v4.3 (#2358)
* Support building with ESP IDF release/v4.3
* Working to fix building for all supported variants (esp32, esp32c3, esp32s2 and esp32s3)
* Update install script to IDF v4.3, apply patches
* Update flashmem, UART to build for esp32s2
* Get builds working for RISCV arch. (esp32c3)
* Use unique SDK project name for each ESP variant

Avoids problems when changing ESP_VARIANT setting
Allows each variant to have customised settings, which we're likely going to need

* Use HAL calls for interrupts
* Revise SPIClass to use more HAL definitions
* Fix Timer for ESP32
* Set SDK compiler optimisation for size
* Configure ADC width according to ADC_WIDTH_MAX
* Patch IDF HAL files to build in C++

+++ Sming/Core/Data/Stream/SectionStream.cpp
@@ -145,7 +145,7 @@ uint16_t SectionStream::readMemoryBlock(char* data, int bufSize)
 		}
 	}
 
-	bufSize = std::min(size_t(bufSize), sect->size - sectionOffset);
+	bufSize = std::min(uint32_t(bufSize), sect->size - sectionOffset);
 
 	return stream->readMemoryBlock(data, bufSize);
 }
